リファクタリング
キャプション追加，変数・関数名の変更
