+++ websocket_server.py
@@ -2,23 +2,28 @@ import asyncio
 import websockets
 import logging
 from pycolor import pycolor,setTermColor
-from rich.console import Console
 from rich.table import Table
-from rich.live import Live
-from rich import print
-from rich.layout import Layout
-from rich.panel import Panel
-import random
 
 class WebsockServ:
+    """
+    Websocket Server 4 GBC-S2Ws
+    """
     def __init__(self,queue:asyncio.Queue) -> None:
+        """
+        Parameters
+        ----------
+        queue : asyncio.Queue
+            Queue from Socket.
+        """
         self.queue = queue
         self.isConnected = False
         self.connections = {}
-        self.console = Console()
         self.isUpdatebleTable = True
 
-    async def sendQueue(self) -> None:
+    async def sendDataFromQueue(self) -> None:
+        """
+        Get Queue from Socket and send assorted data to Websocket. 
+        """
         while 1:
             q:str = await self.queue.get()
             logging.debug(f"< Queue:{q}")
@@ -32,19 +37,20 @@ class WebsockServ:
                 elif q[0] == "p":
                     await self.connections["/playerName"].send(q)
                     await self.connections["/icon"].send(q)
-                # for s in self.connections.copy():
-                #     await s.send(q.encode().decode("unicode-escape"))
-                #     await asyncio.sleep(0)
-                # websockets.broadcast(self.connections,q.encode().decode("unicode-escape"))
             except Exception as e:
                 # logging.error(setTermColor("ブラウザ接続待機中...",pycolor.GREEN))
-                # logging.error(e)
+                logging.debug(e)
                 pass
 
     def createUITable(self) -> Table:
         """
         Create Websocket UI Table.
         When all plugin is connected, self.isAllConnected will be "True" 
+
+        Returns
+        -------
+        printTable : rich.table.Table
+            Created Websocket UI
         """
         printTable = Table(show_header=True,header_style="bold sea_green2")
         printTable.add_column("UI",justify="center")
@@ -52,15 +58,27 @@ class WebsockServ:
         UIs = ["/icon","/playerName","/score","/transition"]
         self.isAllConnected = True
         for UI in UIs:
-            printTable.add_row(f"[{'cyan1' if UI in self.connections.keys() else 'gray66'}]{UI[1:]}[/]","[green]:white_check_mark:[/]" if UI in self.connections.keys() else "[red]:cross_mark:[/]")
+            pathColor = 'cyan1' if UI in self.connections.keys() else 'gray66'
+            connectedEmojiStatus = "[green]:white_check_mark:[/]" if UI in self.connections.keys() else "[red]:cross_mark:[/]"
+            printTable.add_row(f"[{pathColor}]{UI[1:]}[/]",connectedEmojiStatus)
+            # UIのTitle変更用
             if UI not in self.connections.keys():
                 self.isAllConnected = False
+
         return printTable
     
     async def handler(self,websocket) -> None:
+        """
+        Handler for Websocket.(called when new socket connected)
+
+        Parameters
+        ----------
+        websocket : websocket
+            websocket.
+        """
         self.websocket = websocket
         self.connections[websocket.path] = websocket
-        # (self.connections)
+
         logging.info(setTermColor(f"connected:{websocket.path}",pycolor.BLUE))
         #  Table更新
         self.isUpdatebleTable = True
@@ -68,6 +86,7 @@ class WebsockServ:
         async for msg in websocket:
             await websocket.send(msg)
             await asyncio.sleep(0)
+
         try:
             await websocket.wait_closed()
         finally:
@@ -76,11 +95,18 @@ class WebsockServ:
             self.isUpdatebleTable = True
 
     async def websocketMain(self) -> None:
-        logging.info(setTermColor("web start",pycolor.BLUE))
+        """
+        Init function for this class.
+        Run server.
+        """
+        logging.info(setTermColor("Websocket start",pycolor.BLUE))
         async with websockets.serve(self.handler,"localhost",8001):
             await asyncio.Future()#run forever
 
     async def main(self) -> None:
+        """
+        Run server & Send data to Web-browser
+        """
         task = asyncio.create_task(self.websocketMain())
-        task2 = asyncio.create_task(self.sendQueue())
+        task2 = asyncio.create_task(self.sendDataFromQueue())
